middleware: remove unnecessary middleware abstraction
Remove the middleware abstraction and expose http logger directly
